Extract out useful DebugUtils
